% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catRpillar.R
\docType{package}
\name{catRpillar}
\alias{catRpillar}
\title{catRpillar: Functions for handling data from categorisation experiments}
\description{
The catRpillar package provides functions to process categorisation data according to different approaches.
}
\section{Signal Detection Theory}{

The \code{\link{dprime}} function returns sensitivity and criterion (bias) measures according to the Signal Detection Theory.
}

\section{Choice Theory}{

The \code{\link{ChoiceTheory}} function return sensitivity and bias measures according to the Choice Theory.
}

\section{FITA}{

The \code{\link{fita}} and \code{\link{confusion.matrix}} functions offer analysis tools according to the Feature Information Transmission Analysis approach.
}

\author{
Etienne Gaudrain, \email{etienne.gaudrain@cnrs.fr}
}
