% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fita.R
\name{confusion.matrix}
\alias{confusion.matrix}
\title{confusion.matrix}
\usage{
confusion.matrix(d, presented, responded, grouping)
}
\arguments{
\item{d}{A \code{data.frame} containing one trial per row.}

\item{presented}{The name of the factor used for the presented category (as a string).}

\item{responded}{The name of the factor used for the responded category (as a string).}

\item{grouping}{A vector of variable names used to group the matrices by.}
}
\value{
A \code{data.frame} with the following fieds:
\itemize{
\item \code{Presented} containing the name of the presented category
\item \code{Responded} containing the name of the responded category
\item \code{Count} the number of occurences
\item and the factors used for grouping
}
}
\description{
Generates the confusion matrix from a categorisation data frame.
}
\examples{
data_cm = confusion.matrix(dat, 'Presented', 'Responded', c('Condition'))
ggplot(data_cm, aes(x=Presented, y=Responded)) +
geom_point(aes(size=Count, colour=Count)) +
    scale_colour_viridis_c() +
    facet_wrap(~Condition) +
    coord_fixed() +
    guides(size=F) +
    theme(axis.text.x = element_text(angle = 90, hjust=1, vjust=.5))

}
