% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dprime.R
\name{dprime}
\alias{dprime}
\title{dprime}
\usage{
dprime(dat, grouping, presented = "Presented", responded = "Responded")
}
\arguments{
\item{dat}{A \code{data.frame} with the responses in long format (one row per trial).}

\item{grouping}{The factors used to group the d' calculation by (you probably at least want your participant id there). This should be specified as a vector of strings.}

\item{presented}{A string containing the name of the \code{data.frame} field containing the names of the presented categories. (Defaults to "Presented")}

\item{responded}{A string containing the name of the \code{data.frame} field containing the names of the responded categories. (Defaults to "Responded")}
}
\value{
A \code{data.frame} with the grouping factors, the presented factor, and two fields \code{d} and \code{c} that contain the d' and criterion.
}
\description{
Calculate d' for categorisation data. This applies to data where an item pertaining to a category
is presented, and the participant answers a category by choosing amongst a set of choices.
}
