% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fita.R
\name{fita}
\alias{fita}
\title{fita}
\usage{
fita(cm, grouping)
}
\arguments{
\item{cm}{The confusion matrix (see \code{\link{confusion.matrix}}).}

\item{grouping}{A vector of strings containing the names of the variables used to group the FITA by.}
}
\value{
A \code{data.frame} with fields:
\itemize{
\item \code{Hf} The total information entropy.
\item \code{t.abs} The absolute mutual information.
\item \code{t.rel} The relative mutual information (\code{t.abs}/\code{Hf}).
}
}
\description{
\code{fita} calculates the mutual information between presented and responded categories.
}
\references{
Based on Oosthuizen & Hanekom (2016, Speech Comm.) and Miller & Nicely (1955, JASA).
}
\seealso{
\code{\link{confusion.matrix}}
}
